Version 2.1.2
This release is compatible with Clojure 1.0 and
Clojure-Contrib 919 (in old SVN terms). Other
pre-GTIC contrib versions from Github should also
work. For Clojure(-Contrib) head consider using the
bleeding-edge branch from the repository.

The Ivy configuration will fetch an known-to-work
Clojure(-Contrib) version. For manual compilation,
you have to care for that yourself. A compiled jar
is already included in the distribution in the build
subdirectory.

Changes since v2.1.1:

* Fixed a regression of qualified keywords highlighting
* Fixed a infinite loop if there is no prompt in the Repl buffer
* Made compilation of ng client a bit more comfortable
* Added library configuration for OpenSolaris (thanks to Richard Newman)
* Changed normal invokations to ignore keymappings (reported by James Reeves)
* Fixed typo in lispwords: defmethod -> defmulti
* Added letfn to lispwords
* Added c.c.def/defnk to lispwords
* Added #_ and #! as comment indicators to syntax highlighting
  (currently only the #_ is marked as comment, the following form
  is still highlighted normally.)
* Fixed annoying o/O delay bug
* Simplified plugin installation